% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJ_colours.R
\docType{data}
\name{EJ_colours}
\alias{EJ_colours}
\title{Colours based on EJ}
\format{An object of class \code{character} of length 9.}
\usage{
EJ_colours
}
\description{
Colours based on EJ
}
\keyword{datasets}
