% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJ_cols.R
\name{EJ_cols}
\alias{EJ_cols}
\title{Function to extract EJ colors as hex codes}
\usage{
EJ_cols(...)
}
\arguments{
\item{...}{Character names of EJ_colours}
}
\value{
name and hex code for EJ colours
}
\description{
Function to extract EJ colors as hex codes
}
\examples{
EJ_cols("killarney moss","manitoulin sky","niagara cabernet","haliburton green")

}
\author{
Eugene Joh, MPH \email{ejoh@bu.edu}
}
