% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJ_pal.R
\name{EJ_pal}
\alias{EJ_pal}
\title{Return function to interpolate a EJ colour palette}
\usage{
EJ_pal(palette = "primary", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in EJ_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
vector() of selected colours based on palette argument
}
\description{
Return function to interpolate a EJ colour palette
}
\examples{
EJ_pal("accent")(5)

}
\author{
Eugene Joh, MPH \email{ejoh@bu.edu}
}
