% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJ_palettes.R
\docType{data}
\name{EJ_palettes}
\alias{EJ_palettes}
\title{Palettes based on EJ}
\format{An object of class \code{list} of length 4.}
\usage{
EJ_palettes
}
\description{
Palettes based on EJ
}
\keyword{datasets}
