% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_EJ.R
\name{scale_colour_EJ}
\alias{scale_colour_EJ}
\title{Load ggplot2 scale for fill aesthetic options based on EJ colours}
\usage{
scale_colour_EJ(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in EJ_palettes}

\item{discrete}{Boolean indicating whether the discrete variable used}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to discrete_scale() or scale_color_gradientn()}
}
\value{
vector() of selected colours based on palette argument
}
\description{
Load ggplot2 scale for fill aesthetic options based on EJ colours
}
\examples{
require(ggplot2)
data(mtcars)

ggplot(mpg, aes(x=hwy, y=disp, size=hp)) +
  geom_point(aes(col=model) +
  scale_colour_EJ(palette="main", discrete=TRUE, reverse=TRUE)
}
\author{
Eugene Joh, MPH \email{ejoh@bu.edu}
}
